requirements.txt
Flask
Flask-Cors
dotenv
SQLAlchemy
google-genai